Fix reward Bug
