+++ lib/widgets/review/unsolved_quiz_wrapper.dart
@@ -1,14 +1,6 @@
-import 'package:booqs_mobile/data/provider/user.dart';
-import 'package:booqs_mobile/data/provider/todays_answers_count.dart';
-import 'package:booqs_mobile/data/remote/quizzes.dart';
-import 'package:booqs_mobile/models/answer_creator.dart';
 import 'package:booqs_mobile/models/drill.dart';
 import 'package:booqs_mobile/models/quiz.dart';
 import 'package:booqs_mobile/models/review.dart';
-import 'package:booqs_mobile/models/user.dart';
-import 'package:booqs_mobile/notifications/answer.dart';
-import 'package:booqs_mobile/utils/answer/answer_feeback.dart';
-import 'package:booqs_mobile/utils/answer/answer_reward.dart';
 import 'package:booqs_mobile/widgets/quiz/answer_part.dart';
 import 'package:booqs_mobile/widgets/quiz/question_part.dart';
 import 'package:booqs_mobile/widgets/quiz/unsolved_content.dart';
@@ -44,42 +36,13 @@ class ReviewUnsolvedQuizWrapper extends ConsumerWidget {
     final answer = QuizAnswerPart(quiz: quiz);
     final footer = QuizUnsolvedFooter(quiz: quiz, review: review);
 
-    // サーバーからのレスポンスをもとにProviderを更新する
-    void _updateProviders(resMap) {
-      // ユーザー情報を更新する
-      final User user = User.fromJson(resMap['user']);
-      ref.read(currentUserProvider.notifier).state = user;
-      ref.read(todaysAnswersCountProvider.notifier).state =
-          user.todaysAnswerHistoriesCount;
-    }
-
-    // 解答をサーバーへリクエストして、結果に応じて報酬を表示する。
-    Future<void> _requestReview(notification) async {
-      Map? resMap = await RemoteQuizzes.answer(context, notification, 'review');
-      if (resMap == null) return;
-      _updateProviders(resMap);
-      final AnswerCreator answerCreator =
-          AnswerCreator.fromJson(resMap['answer_creator']);
-      // awaitをつけるとAnswerRewardを表示が重なった時にLooking up a deactivated widget's ancestorエラーが起きる
-      AnswerFeedback.call(answerCreator);
-      // 報酬を表示する
-      await AnswerReward.call(answerCreator);
-    }
-
-    return NotificationListener<AnswerNotification>(
-      onNotification: (notification) {
-        _requestReview(notification);
-        // trueを返すことで通知がこれ以上遡らない
-        return true;
-      },
-      // 解答時に問題をフェイドアウトする際にリビルドされてレイアウトが崩れるのを防ぐために、外部からwidgetを渡す
-      // ref: https://qiita.com/chooyan_eng/items/ec11f6dcf714f7a2fa3d
-      child: QuizUnsolvedContent(
-          quiz: quiz,
-          header: header,
-          question: question,
-          answer: answer,
-          footer: footer),
-    );
+    // 解答時に問題をフェイドアウトする際にリビルドされてレイアウトが崩れるのを防ぐために、外部からwidgetを渡す
+    // ref: https://qiita.com/chooyan_eng/items/ec11f6dcf714f7a2fa3d
+    return QuizUnsolvedContent(
+        quiz: quiz,
+        header: header,
+        question: question,
+        answer: answer,
+        footer: footer);
   }
 }
